Add a hard limit on the size of the widget images.
There is a limit on the size of the data transfered by binder.
For now, we just add a hard limit (360 pixel) to ensure the widget's image
can be passed by binder.

Also adjust the size of widget to make it looks better.

Fix a bug in DecodeUtils which cause OOM for a image in size 12200x1920.
In that case, we should generate a screen nail of size 640x101 instead of
4066x640.

Change-Id: Ia8227d8e5368471fe7af94bf164d67017aa321fa
fix: 5273271