Pause the playback of video if the activity is resumed after 3 mins.
fix: 5260232

Also keep the current playback position when the activity is resumed.
Originally, we use the bookmark to keep the position. However, the bookmark
will be ignored if

1.) The video duration is less than 2 mins, 2.) The video
is just started (with 30sec), or 3.) The video is approching the end.

I think resume should not limited by the above rules. And we did get some
reports on this. (We mark them as WORK AS INTENTDED).

Change-Id: Ib9ac359d7ccb6a0278777b0892f0821e16bbb745