Improve screen on/off logic in slideshow.
When used as a screensaver or otherwise, the slideshow
should not keep the screen on if the device is not plugged
in. It should also allow the screen to lock while the
slideshow is playing (otherwise a device left in this state
may be left unsecured for long periods by accident).
Additionally, when used as a screensaver, it must dismiss
the keyguard in order to be shown on a secured device.

Bug: 5660833
Change-Id: I1bafc50127aaa7ae2b67235d30736faa606c5ac1