Fix 5948189 BUG BASH: action bar flashes during multi-select
This CL fixes the flashing by disabling MenuItem directly instead of setting null intent
to ShareActionProvider, so there won't be any "icons shrinking" effect that causes flashing.

The rest of the change is code refactoring. More comments are added and the final UI operations
are merged into one Runnable object.

Change-Id: I73843c1df6b7a9675176e1b3ae6a35cc1cb715fa