Fix ANR in PhotoPage.
This bugs happens because we open too many PhotoPage(s) (and onSingleTapUp is a delayed event).
However, each of the PhotoPage need a Thread to run tile decoder. When we close one of the
PhotoPage, it will wait the tile decoder to get finished first. However, the title decoder may
still waiting in the queue and never got a chance to run.

Change-Id: I113d1150729892edb4fe36bc5a1dc131db300476
fix: 6319833