Improve image size change handling.
- When the camera relative frame is set, we delay the change for camera
rectange until the next setImageSize call. This avoids the unwanted
intermediate state if we change them separately.

- Don't draw other screennails when the camera is in full screen. This
avoids showing other screennails when the image size changes.

- When the aspect ratio of the image changes, we assume the view angle
of the longer side doesn't change (so the aspect ratio change is because
the view angle of the shorter side changes). This matches what camera
preview does.

Bug: 6566612

Change-Id: I7603416f31c96ba77c96cdc2a3d0b79f8921c491