Add the UI elements for video trim activity.
1. Refactored the classes to support both the video play and trim.

Now the common code for controller overlay is in CommonControllerOverlay,
which is basically the same as the original MovieControllerOverlay without
animation. Because in trimming, we don't want to hide all the controller.
The specific animations are implemented in the sub-classes which are
TrimControllerOverlay and MovieControllerOverlay.

At the same time, TrimTimeBar extended TimeBar to support the extra
trimming start scrubber and end scrubber.

The interface between the timebar and the controllerOverlay are kept almost
the same way, except adding the trimming info when necessary.

With all these, the activity of TrimVideo now relied on the TrimTimeBar
and TrimControllerOverlay. Similarily, the MovieActivity relied on the TimeBar
and MovieControllerOverlay.

2. Hook the TrimVideo activity with the trim menu.

Note that the icons are temporary for now, still need UX input on that.

bug:7093055

Change-Id: Ib9bfbc090106744a569fce4c451ddffc0a2c699b