Reskin: Correctly position rotated thumbs in album listing
Bug: 7067732
Bug: 7050303
Changing the slots to be non-square surfaced a small math
error, and also raised the question of where exactly to
render the texture. The solution that makes the most sense
given the current design is to fit the thumbnail (which is
always square in our case) in the largest square that fits
inside the slot, and aligned to the top.

Change-Id: I98d94ad9255399aa4f1e5c403ff81811674bb264