Fix pinch out gesture for film strip mode on gingerbread.
On crespo running Android 2.3.6, a pinch out gesture results in the following
call sequence: onDown(), onUp() and then onSingleTapUp(). The correct sequence
for a single-tap-up gesture should be: onDown(), onSingleTapUp() and onUp().
The call sequence for a pinch out gesture in JB is: onDown() and then onUp().
Base on these observations, a condition is added in onSingleTapUp() to filter
out the false alarm where onSingleTapUp() is called within a pinch out gesture.

Bug: 7023717
Change-Id: I010b87f95c3197a299a7d907ce32e93a1d5e958f