@#!$# Bug #7275145
Removing the window background when it's not visible is a good idea
to save on performance. Removing *all* the backgrounds is just a silly
idea that leads to weird looking results that drive graphics engineers
mad for a day until they figure out that everything is working properly
and that the app is simply asking the system to draw whatever garbage
is left in the GPU's graphics buffers.

Change-Id: Icd12cd4ec6fc313f80acfd8deb7133a17ea11302