DO NOT MERGE: Do not set the background to null for video.
Bug #7275145
Bug 7405595

Removing the window background when it's not visible is a good idea
to save on performance. Removing *all* the backgrounds is just a silly
idea that leads to weird looking results that drive graphics engineers
mad for a day until they figure out that everything is working properly
and that the app is simply asking the system to draw whatever garbage
is left in the GPU's graphics buffers.

Cherry pick the Romain's change from jb-mr1-dev:
https://googleplex-android-review.googlesource.com/#/c/234571/

Change-Id: Ief88821319989add5fda57ae7b69cdb01f6ccdf4