Set the thumbnail/screennail sizes based on screensize
Bug: 7232758
Prior to this CL, the thumbnail and screennail sizes were
hardcoded which resulted in mediocre image quality when
using the Gallery in general and the camera filmstrip.
Now, we set the size of these items based on the resolution
of the display. Because high-resolution devices tend to
have higher-performance graphics to drive their large
displays, we expect that this will not introduce any
significant performance impact.

Change-Id: I90e08f7aca670e275c913fd9e21e4459c8bf9fcf