Add relativePath field into photo widget db
Bug: 7481248

The bug is caused by the change of external storage going from JB to JBMR1.
In light of this change, a new field has been added to the photo widget app
database to store the relative path for the local album. With the relative paths
stored, the widget app should be more resilient to future storage path changes.

Change-Id: Ia2497b882ae67178fa0632f23e07673b82d3a942