Don't call requestLayout from onLayout
Doing a layout within a layout can lead to various problems, so it should
be avoided. There's some logging that warns of this, which got triggered
by CommonControllerOverlay.
I don't know what problem this code was originally trying to solve, but it
doesn't seem to be necessary anymore. The line in question was originally
introduced in change 209a9163d4e8cee0bfe162ae598ef40e6051479c, in Oct 2011.

Change-Id: I1402a842df59df31b69ad981337ea1c8630103df