Enforce bmp type in pool, handle MTP image decode failure
- GalleryBitmapPool didn't reject bitmap types other than
ARGB_8888 which could lead to problems when a different type
is used for recycling.
- BitmapFactory throws an exception rather than returning null
when image decoding fails and an existing bitmap was supplied
for recycling, even if the failure was not caused by the use
of that bitmap. When decoding things from an MTP device, we
need to handle this since unsupported formats may be returned.

Change-Id: I8cc8aa46f5a741b360e806814991f74a131e9039