add up button support for filmstrip
This CL adds up affordance handling to CameraActivity. There was no working call to reset the filmstrip back to the camera, so this CL adds a controller which then calls reload, which is itself fixed to work as desired. Note that during testing I identified issue 10563392, I've decided to file that separately since it is orthogonal to up button function, but it does affect the overall experience and should be fixed.

Bug: 10414498
Change-Id: I1ee914dc1d264adb1fb5a363c2d6174070cd7b71