Add back button handlng to CameraActivity.
This CL first detects whether user is in filmstrip mode i.e. not in Camera,
if so brings user back to the first item which is the Camera. If in Camera,
we route back handling to the current Camera module which enables already-
written module handling for back button e.g. backing out of pie menus.

Bug: 8781981
Change-Id: I18377ea5fbdcd8bd8c6fc3af9cb5322d5f5a889a