camera: Invoke camera UI functions only after initialization.
When switching from Panorama mode to Camera mode, the CameraActivity
invokes two functions openModule and onOrientationChanged.
openModule initializes PhotoModule. As part of the PhotoModule
initialization, the PhotoUI and PhotoMenu are initialized.
But this is acheived by posting a message to the handler. When the
message is read by the Handler, the PhotoUI's onCameraOpen is invoked
which initializes the PhotoMenu. But sometimes before this message
is read by the Handler, the onOrientationChanged is invoked, which
invokes couple of functions from PhotoUI class acting on the PhotoMenu.
But since the PhotoMenu is still not initialized, it throws up error in
the form of NullPointerExceptions.
Fix this by checking if the PhotoUI and PhotoMenu have been initialized.
If they have not yet been initialized, send a message to the Handler to
invoke these PhotoUI functionalities after the PhotoUI has been
initialized completely.

(cherrypicked from commit 774659c9f580b90a15a896881376075bb1506a05 )

Change-Id: Idb55e4f5ba29523c273ff66f11149ff271c490a4
(cherry picked from commit 42f1ffca2e43c4561f5bc3534a966c4bbe6cbef8)
(cherry picked from commit 8df298b693932da2c06dbfd40ad5bbe57dc96308)