Fix build break.
ag/369626/ added a cannot_load_image string which was translated in ag/370960/
However, this translation resulted in a duplicate string since this string
already existed in translated versions of filtershow_strings.xml which was
deleted in ag/339346 but the translated versions weren't deleted. Deleting all
the translated versions of that file caused a problem because ag/362583/ added
source code that depended on a setting_wallpaper string which only existed in
the translated string files.

This change deletes the translated filtershow_strings.xml and adds the
untranslated setting_wallpaper string.

Change-Id: I1b8ac8c854c6f06abde9a71db2105c4d2dd44934