Consider layout width for pie menu arc radius
Pie renderer arc radius was always constant and read from resource
dimensions, irrespective of the screen width. Hence when there are
more options, some of the options were getting rendered beyond the
screen.
In this code the value from dimensions is saved to a separate
variable. During each layout change maximum value which can be
rendered is set as arc radius.

Change-Id: I893685d0e57c839ccbb96826c01208fa581fa7cc
CRs-Fixed: 568052