Panorama: Restart take picture call after layout changes.
In reverse landscape and reverse portrait panorama modes, layout
change callback is received when shutter key is pressed for the
first time. It looks like the layout change call is because of
navigation panel getting rendered on top of the screen. Due to
this camera application resets the preview window dimensions and
changes the status from capture to view finder.
This change restart the start picture call after layout changes.
Its done through hanlder because reset configMosaicPreview() posts
a message to reset on a handler.

Change-Id: Ic42a6929198eff52d659468f1df7170bdbdf6a38
CRs-Fixed: 573570