Camcorder default video quality preference updated as per UI.
Video quality preference was not initialized to the default value.
But it was read as null and set to maximum available quality. So
the actual value set to camera is different from the one user see
in the settings list.
This change will set the default quality to the preference if its
supported.

Change-Id: Ie219e0965a6ac06e62f32cc6c52fdf255a07adee
CRs-Fixed: 565374