Update storage hint when switched back to full screen preview.
If user swipe right and delete some images to empty space. Or pause
the activity, delete some data to free up memory also the storage
hint was not getting dismissed. Because avaiable storage space was
not checked in all these cases.
This change makes sure the storage availability is checked and UI is
being updated.

Change-Id: I29a0471d596945ae370b21a61e3ad1f995dced25
CRs-Fixed: 573549