Add a capture-session API to be used by all modules.
  Bug: 11747284

This refactors the way we think of sessions. Instead of
having multiple session and notification APIs being used
by different modules, we create on central capture session
API to be used by all modules.

This also adds a first implementation of a Memory API which
tells modules about the current memory situation.

Change-Id: I7f030e32fc2f70f4007825ba7bbbdce9521a2bd9