Do not add screen hint dialog when window is not available.
SD card screen hint is directly added to activity window. In the
error case its trying to add the screen hint dialog even before the
activity window is available. This change check for the availability
through window focus. In photo module update hint is called through
handler to make sure it adds the dialog in UI thread, which was
missing in panorama modo. This change as photomodule kind of UI
updation calls in Panorama mode.

Change-Id: I5cad2e86a23aceff290ef8d4ef89dbfdaaeb683d
CRs-Fixed: 573549