Camera: Revering changes of clear preview frame during activity pause.
In camera application, the preview frame is rendered on the surface
texture. When user pause, this surface texture is retained by the
activity. Upon user resume, it will be destroyed and recreated so that new
preview frames can be rendered on it. During this period, a flickering
impact is seen on UI.

Change-Id: I5728d5d7f28ec45121f375928bf2c9effd9bd620
CRs-fixed: 583474