add a11y announcement when multi toggle button changes
addressing TODO with a manual announcement whenever the content
description of a button changes. Does checking of existing content
description to reduce chattiness when options are set by the system
e.g. on a resolution or orientation change. Still a bit of chattiness
on startup if the app is setting a button to a nonstandard value, but
I think its acceptable, it's giving user useful information about
nonstandard options on startup.

Bug: 10745077
Change-Id: I81089b2f4f7f14b90c4436199895a0f67075c3d3