Force chooser for sharing actions
User should not be able to select 'always' in the Camera filmstrip, so this
uses the system chooser. This does not do the same for panoramas. This is a
simple implementation, no querying for counts of intent receivers, but having
just one for images or videos is a corner case, showing the list is safe.

Bug: 13229214
Change-Id: I0b7a8327b157828e03edf65579318b758e1873e8