Remove onPreviewReadyToStart from PhotoModule
Removing this call results in no jank or loss of function on tested
devices, and resolves an issue on certain devices where starting the
callback prior to the preview results in a permanent mode cover. This
issue was not observed in any other mode, i.e. video is working on
warm start for all tested devices before and after this change.

Bug: 13078926
Change-Id: I7a63011dd96d40466d6bc3bb7f6867552a2b0dc8