Fix HDR+ landscape transform
Some logic in resetting translation overlooked the fact that HDR+ has
rotation baked in the transform matrix because camera2 api requires
that surface texture gets rotated by app for device rotation. As a
result, simply setting the translation to 0 for transform matrix when
there is rotation set on the matrix as well offsets the preview off
the screen.

Bug: 13367662
Change-Id: I0529fe4913b55aebc058c93a901b6689da1208ba