Hide bottom bar of Camera when filmstrip is visible
This CL hides the bottom bar, i.e. capture button and onscreen indicators,
when the filmstrip is visible. We hide a wrapper; modules may independently
control the visibility of the items in the wrapper e.g. Photo Sphere turns
the actual buttons invisible even if the wrapper becomes visible.

This change takes care of the most egregious issues with widgets visible to
TalkBack under the filmstrip, but there are still others e.g. 'align to start'
for Photo Sphere, can be navigated to. I'll file a separate bug to pursue
a more aggressive wrapper visibility toggle which turns off the whole camera
preview, mode switcher, etc.

Bug: 13644558
Bug: 13589550
Change-Id: I1cf7e9079210c85f29b3e35e0f7dbf176b86976e