Camera2: Don't report incorrect supported picture formats
If camera HAL does not return valid picture formats do not show wrong defaults

Currently it falls back to jpeg and raw to choose for the user. All our devices which do not have a proper
return of picture format only support jpg. So remove the entry to choose in this case completely

Change-Id: Ib4e4094ba0a57e7751f6568f6ec3c3f1dfc1ffb8