Avoid a race condition causing dup'd images.
When a session is finished, it is written to the
media store.

We now have a listener looking for new images in
the media store.  If that happens to fire before
the session manager has updated the filmstrip,
the photo ends up being added twice.

Now, the listener checks to see if it is a photo
already being handled by a session

Bug: 13644124

Change-Id: I8a3c4a161ad0c326d9295f6714f6c6cbbcc6764e