Make the camera id setting have static possible values.
This is important because the dynamic hook for querying the
supported number of cameras is currently only based on the camera 1 api.

We want to make sure we can set the enabled/disabled state of back
and front facing camera in the mode options, so it's necessasry to
always have both possible values.

Bug: 13769609
Change-Id: I9d526315dbe15e008542832f9b2c4c9d23d34021