Replace isDebugging compiletime logging override with runtime property
This CL accomplishes the same goal as the compile-time isDebuggable,
with a runtime property check. This is useful so that we opt in to
logging everything at a given level with one setprop line.

Note that either with this new runtime property or with isDebuggable,
the Log proxy wouldn't log if a particular Configuration value is false,
since that level is instead routed to SILENT_LOGGER.

Bug: 13737123
Change-Id: I78ce629620f420a08a2cc013865847674941fb9b