Camera: Block HFR option for Video HDR
Video HDR can’t work with HFR and that caused AEC behaviour change.

Fix Made: User'll be notified that HFR/HSR selection is not allowed
for HDR mode. After that, HFR selection will be set to off and the
option will be disabled. It'll not be vice versa as HDR has higher
priority than HFR.

[mikeioannina]: Removed HSR code

Change-Id: I4b0f8362667a53e46a83223cab77b7248e1410e2
CRs-Fixed: 652078