Fix storage warning hint crash
We rarely hit this warning, but indications are that it crashes 100%. This is an old
class, not even sure why we're using it vs a Toast, but for now fixing the existing
implementation. OnScreenHint accepted a context but attempted to add views which
requires an Activity. Likely started failing when we used to in practice pass Activity
contexts but converted to Application contexts. I have converted OnScreenHint to require
an Activity. We have one device afflicted by this 100% which I used to test, and it now
works with this CL and shows a storage warning.

Bug: 14134552
Change-Id: I1d8a7a7dce799c81c821b9974e8d8d62c6daab39