Maximize space for Large Text in aspect ratio dialog.
This evenly places the vertical element in portrait mode of
the aspect ratio dialog.  This gives the maximum space for
translations.  If the translations don't fit, it is because
we are taking up all of the space on the screen.

Tested:
French, German, Turkish, English
In both Large Text and normal Text mode.

Bug: 14993598
Change-Id: I76ce1c00bbbb05733211d6e498a8bcf44777933a