Rename HDR+ torch to flash
For consistency, and to better reflect the meaning of the
flash button in HDR+ mode, change the naming from 'torch'
to 'flash'.

Even though flash in HDR+ mode is implemented by turning
the flash on in torch mode for a burst of frames, HDR+ flash
is closer in meaning to a regular photographic flash than
a flashlight/torch.

Bug: 15133661

Change-Id: I3ac4bb3dac455be1341575966910c2a86e29644b