Camera2: Still popup tips when change scene to other after turn off HDR
The toast popup or not depend on the value of hdr and scene mode,
but now the value mHdrOn status is not uniform with the actual hdr
value, which lead to popup toast incorrect.

Get the value of scene mode or hdr from the preference directly rather
than a new variable. So the mHdrOn is not used and delete it.

CRs-Fixed: 674521

Change-Id: I1c6926e0ba295e9869109b7412a7413323d3bd1c