Hard reset HDR off if HDR+ is supported.
HDR and HDR+ should not be activatable on the same phone.
Our app depends on this assumption, so in case the app is killed
in some way to prevents us from going to gms core to get the
right information about hdr/hdr+ support, force HDR off if HDR+
at any point is supported.

Bug: 15310749
Change-Id: I450671d1218679ac8d413a01c5857030ba6ea32a