Modify Camera settings scope enforcement from crash to warning
Several bugs have appeared particularly for flash settings, where the app
is attempting to get a scope per the new SettingsManager, and the camera
is already closed meaning there is no valid scope. With this change, such
situations will be logged with call stack traces, so if we see deviant
user behavior we'll know the source. Monkeys should no longer trip settings
crashes with this change.

Bug: 16483952
Bug: 16491601
Change-Id: I26487bc58909a05183563f59ae3cfba571f2504e