Upgrade mode index settings after mode indices were cleaned
Because we removed a mode index from the camera_mode_* sequence in res/values/arrays.xml,
upgrades were found to crash when settings referred to the old, now nonexistent mode value.
This CL preserves the arrays cleanup and adds a settings upgrade to convert modes to the
new value. Tested and works properly with mode preservation.

Bug: 16467318
Change-Id: I1352a940285687af38290250d7a06b5b7c421338