Restore grid initialization logic to fix grid display on startup
The large settings refactor in ag/484817 not only changed settings
calls in ButtonManager.initializeGridLinesButton, it altered the
order of calls to create the button callback and to set the button
state. The prior implementation and this CL creates the callback,
and then sets state so that the callback, which controls display,
runs upon initial state setup of the button. Otherwise the grid
is not shown correctly on module startup.

Bug: 16687954
Change-Id: I482906d2d54a8545066dded3209b55ae4ad48bdb