Refactor Upgrade/UpgradeAosp into SettingsUpgrader
This CL does three things:
- removes the Upgrade/UpgradeAosp classes which were bunches of statics,
in favor of a SettingsUpgrader abstract class that can be overridden
- moves 'upgrade to strings' hack logic into AppUpgrader so it only is
checked and run once, not for any of the former UpgradeSteps. logic used
to be in Upgrade and run regardless of key, when it was using detection
only meant for KEY.KEY_UPGRADE_VERSION
- adds existence checks for all hack in-place key string upgrades so they
are no-ops if the key was not persisted in the first place.

Bug: 16573087
Change-Id: Ife42f7aa4a7611ba2d143410a3450057b7464901