Move preference subscreens to new Activities from Dialogs
Dialogs are hard to override for action bar settings. The documented
way to do navigation up is to receive menu events. So we no longer
use Dialogs, instead here we set up intents back to the same Activity,
and add an extra system to designate which subscreen to load. This uses
activity stack to do subscreen navigation, and then all screens hit the
same menu handler in the Activity which properly receives up navigation
signals via 'home'. This is not awesome but a better hack than the
Dialog one which depends on Action Bar internals; this approach only
uses public APIs.

Bug: 15756177
Change-Id: Iae2bd254e9cd5dde1d32d0997e5610f359662467