Camera2: Do not restart preview to update flip params during orientation change.
When there was a orientation change flip paramter had to be changed.
Currently all parameters are set again and preview is being restarted.
so preview will not be smooth when user rotae the device.

This change makes sure only necessary params are updated and
preview will not be restarted.

CRs-Fixed: 710129

Change-Id: I564431ef8cedaf68c9d12e2a84a3a430c845ab6d