TextureViewHelper workaround for SurfaceTexture/CameraProxy race
CameraController#getCurrentCameraId() now returns the ID of the requested camera
device if called after #requestCamera() but before the camera device has
finished opening. Assuming the most recent request is for the desired API level
and device, this resolves an issue where no transformation matrix was applied on
returning from the settings activity.

Bug: 17140996
Change-Id: Ideaec0072acd8aab77b98b213f260907f84270de