Set video quality pref default values at startup.
The problem of the old approach is the default values never got set
until switching on video module.

The caveat of the this CL is that the default value tweak is only applied
for Shamu. Rather than basing on device model string, detecting
CamcorderProfile capability is a better way to get this job done. However,
|CamcorderProfile.hasProfile| needs camera id info so I'd argue this CL
is good enough for the Shamu/L release.

Bug: 17060480
Change-Id: Ia53ff6aecad44fe7a32257fa1415163f167c09bf