Remove jank on camera options switching
There are two main causes of jank on encountered on options switching:

1. When requesting camera in requestCamera, a portability refactor
introduced a regression where we blocked in the main thread to close
the camera. This main thread blockage is unacceptable as it caused a
preview freeze in addition to the freeze we were trying to create during
the switch, leading to a two-frame preview freeze.

2. The mode options animations which delay the button state transition,
also resulted in preview stutters. Unsure of cause but reverting first
and will assess fixes and reintroduce.

Bug: 17414652
Change-Id: Ib72e331f1bbb5d3b608d8bdc5656c2ff2836ed42