Fix CameraSettingsActivityHelper
Previous hack to allow activity up button to work, interfered with
custom preference merging we use to conditonally add items to
Settings. Now fixed by allowing internal control of whether the
full preference tree is returned by the fragment or just the
desired subtree, so that the Helper can access the full tree.

Bug: 17534830
Change-Id: I11510f365b9e212f07009b2233bb5d8036228774