Re-run settings files migrations
Increment version number for upgrade and re-run copying. This will
be a no-op for a first time glacier upgrader, copying runs in the
same sequence for this case. For alpha users who hit crashes, this
will re-run the copy, potentially overwriting some values but
ensuring all values are valid. Tested both N4 and N5 upgrade
scenarios that had previously seen issues, this addresses issues
without the need to clear data.

Bug: 17875895
Change-Id: I37c469c1b03151d03ee0d32ce0ba62983af32a36