Preview vs Video Camera Resolution Mismatch
  Bug: 17458889

This bug is programmatic and reproduces on all devices.   Basically,
the current app defaults to the highest resolution preview that
matches the aspect ratio of the recording, even if the recording
video resolution is lower.  Since it is more important to be WYSIWYG,
in the cases where the preview show at the same resolution of
the recording video resolution, we should preview at same resolution
for truest (but not necessarily best) quality that most faithfully
represents the content that is to be recorded under the constraint
of the UI resolution.

Change-Id: I0ec5a76834cbffa59f0e144f1fe271de25e92523