Fix for crash when PhotoApp deletes last Cam pic
    Bug: 17496363

This crash occurs when there is only one picture in the Camera
filmstrip.  When you switch out of the Camera App into the
Photo App and delete it, and then switch back, the Filmstrip
attempts to draw a removal animation with invalid data, and
crashes via trying to call functions on null object.  This fix
is simple and catches the null object and does NOT attempt to
draw the invalid object.

The larger issue may be that the Camera App is buffering/caching
invalid or stale data that it is illegally trying to act on.
I've left a warning message in the code to warn on any bad
behavior, but we should catch the invariant that's being violated
and fix the deeper problem.

Change-Id: I2487f753b5c7bb57e34c65d687a2236fa3a23f9f