Secure Camera App survives phone unlock screen
   Bug: 17883828

There is UI path to start the secure camera app from the lock
screen, and then unlock the phone and find the secure camera
app on the task manager.
This live instance of secure camera app is confusing to the
user who is expecting to interact with a full camera app.

This fix kills the secure camera app when the phone is unlocked
and ensures that the user cannot easily restart the secure
camera app from the currently running tasks in task manager.

Change-Id: Ia6a6d077c460e66807d4be9836fe192184aa8eaf