Button fix for torch when no flash available in hw
    Bug: 17952997

The logic to show and disable the flash button was NOT paired
with the torch button and sometimes allowed the enabled
torch button to be shown, even the flash button was disabled.
A bit confusing, since the torch and flash button look exactly
the same.

Torch button logic to show/hide and enable/disable now follows the flash
logic.

Change-Id: Ie14b07815b6d5e1f8d7ce792ef9ee36dfdf8e3ad