Catch exceptions in settings remove helpers
Saw just one case of this and my hypothesis is that it was due to
another now-fixed exception resulting in a partial upgrade. But
in no case should these remove issues fail, so fixing to remove and
return the default if the underlying value is not readable. There
is some possibility for behavior error in this case since perhaps e.g.
a String-based 'true' was written and we upgrade instead to a default
'false', but allowing users to enter camera is better and I've audited
all uses of these, behavior should be fine/recoverable vs and
preferable to startup crash.

Bug: 18042161
Change-Id: If2453a819311bfd9233fdf65d7a97d8a243b393a