Align id selection from QueryTask to sort order from Media query
In QueryTask we pick the id from the first result and consider it
the max id, yet we query ordered by date taken. On most phones
they're the same but by correcting the query to satisfy the
assumption in QueryTask, we remove costly scenarios where most
photos are reprocessed in LoadNewPhotosTask. One line fix in
LocalMediaData with more verbose logging added in CameraDataAdapter.

Bug: 17437217
Change-Id: Ia643f38d1bfb77ff0beaba989f0effe9a4dec31c