Hook burst controller to the shutter button long press.
Hooks up the shutter button long press to start the burst.
Introduces a BurstManager to interact with burst.
Provides an implementation of BurstManager that listens to burst and
saves results when burst is complete.

The burst eviction handler is now installed on the ring buffer
on shutter button press and uninstalled on shutter button release.
The ring buffer is cleared before starting and after completion of the
burst.

Also provides a stub implementation of the BurstController that has a
static method which controls if burst is enabled or not.

Bug: 18332704

Change-Id: I1098937bf348af7acbf55da1a5eeb423c30fb901