Second Pass at Integration of ImageBackend
Implementation of single service to centralize task submission and
execution of Android Image dependent tasks to minimize the locking
of system resources and to prioritize processing with thread
priority and enforced temporal dependence.   This implementation
is ready to support the plumbing for Lucky Shot.   Currently, the
object are isolated and not wired into the Camera App itself.
The integration of these objects will occur in a subsequent CL.
The unit tests for these object are contained in a separate,
but related CL: ag/602679 .

Change-Id: Ia51d46651ed02ea2d5a014943d17cadb56e9006f