Passing an intermediate preview to the session.
Generalized the image thumbnail preview task to create artifacts of
multiple shape.  Added this task before the full-sized JPEG decode.
Added a field in the TaskInfo to specify the purpose of the imaging
artifacts on the listener, since the processing of an image may result
in multiple uncompressed/compress artifacts and need extra
explicit identification to be properly routed by listener.  Changed the
default behavior of the ImageBackend to produce an intermediate-sized
preview on job submission.  Routed this intermediate-sized preview to
the CaptureSession for an earlier Filmstrip update.

Bug: 18841584
Change-Id: I8db2d3a7c30cddd1e0e9ddf0709b7867f841b056