Remove exception when releasing a requested, not yet opened camera
From monkey tests, we see a spike in an exception from new device
which is likely due to slow close performance. If we camera B is
open, and we request camera A, B will be closed then A opened. If
we release A during this process but B is not yet closed, don't
crash, just let the pending underlying open of A happen and release
the request. Preserve the exception for application errors where
truly unrequested cameras are released.

Bug: 17975345
Change-Id: I9aa3b0575e130f516cea6a3df120e4812df55bce