Fix for JPEG Compression w/ Unknown Orientation
TaskCompressImageToJpeg blindly passed the orientation from the camera to
JpegUtilNative.compressJpegFromYUV420Image, which then threw a run-time
exception.  Behavior for TaskCompressImageToJpeg with an image of
UNKNOWN orientation is to treat it as CLOCKWISE_0 for purposes of
JPEG compression.  Instead of crashing, the camera now receives a JPEG
of the dark image of one's tabletop.

Bug: 19000723
Change-Id: I412fb6fed5d80035b180db7f6e8504d6d36b4fcc