Image Backend Integration w/ Simple Camera
Wiring for cameras that only deliver compressed JPEG artifacts.
This CL was tested by enabling it on the Nexus 5.
Currently, the code in this CL is NOT exposed in any known
Nexus Camera model.  The two unexposed bugs are as follows: an camera
initialization bug on Nexus 4, and a cross-device functional issue
with the SimpleCameraOneFactory object in that it doesn't request
a JPEG compression artifact from the HAL with zero orientation.

Bug: 18908116
Change-Id: If7080c4e8e52a329fb9a6fed52f7c31541758afb